// types.ts
export type Micronutrient = 'Vitamina C' | 'Cálcio' | 'Ferro' | 'Vitamina D' | 'Vitamina A' | 'Potássio' | 'Magnésio';

export type MicronutrientGoal = {
    [key in Micronutrient]?: {
        amount: number;
        unit: string;
    };
};